+++ src/Parser/EditorConfigDocument.cs
@@ -29,7 +29,7 @@ namespace EditorConfig
         public List<Property> Properties { get; private set; } = new List<Property>();
 
         /// <summary>A list of all the error suppressions in the document.</summary>
-        public IEnumerable<string> Suppressions { get; private set; } = new List<string>();
+        public List<string> Suppressions { get; private set; } = new List<string>();
 
         /// <summary>The root property of the document if one is specified</summary>
         public Property Root