+++ src/Parser/ItemType.cs
@@ -1,0 +1,13 @@
+﻿namespace EditorConfig
+{
+    public enum ItemType
+    {
+        Comment,
+        Section,
+        Keyword,
+        Value,
+        Severity,
+        Suppression,
+        Unknown
+    }
+}