+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - [x] Validate spaces in globbing patterns
 - [x] Implement error suppression
 - [x] Add suppressions from light bulbs
-- [ ] Validate suppression syntax
+- [x] Intellisense for error codes in suppression
+- [x] Hover tooltips for suppressions
+- [ ] Validate error codes in suppression comment
 - [ ] Add item template for .editorconfig files
 
 Features that have a checkmark are complete and available for